+++ criu/include/mem.h
@@ -50,9 +50,4 @@ int prepare_vmas(struct pstree_item *t, struct task_restore_args *ta);
 int unmap_guard_pages(struct pstree_item *t);
 int prepare_mappings(struct pstree_item *t);
 bool should_dump_page(VmaEntry *vmae, u64 pme);
-
-int init_pidfd_store_sk(pid_t pid, int fd);
-int init_pidfd_store_hash(void);
-void free_pidfd_store(void);
-int send_pidfd_entry(pid_t pid);
 #endif /* __CR_MEM_H__ */
